Pre release

See merge request auto/order-center!24

+++ account-all/account-renter-wz-deposit/src/main/java/com/atzuche/order/entity/AccountRenterWzDepositEntity.java
@@ -23,7 +23,7 @@ public class AccountRenterWzDepositEntity implements Serializable {
 	/**
 	 * 主订单号
 	 */
-	private Integer orderNo;
+	private Long orderNo;
 	/**
 	 * 会员号
 	 */
