+++ order-core-all/renter-all/renter-order/src/main/java/com/atzuche/order/entity/RenterOrderChangeApplyEntity.java
@@ -23,7 +23,7 @@ public class RenterOrderChangeApplyEntity implements Serializable {
 	/**
 	 * 主订单号
 	 */
-	private Integer orderNo;
+	private Long orderNo;
 	/**
 	 * 租客子订单号
 	 */
